CRM_Mosaico_Page_Editor - Simplfying rendering lifecycle
The normal Civi rendering pipeline is designed to embed within a CMS page
and cooperate with the CMS on resource management.  However, the Mosaico
page is different -- it occupies the entire UI. This requires a work-around
to opt-out of the normal CMS/Civi services.

The `snippet=2` technique comes close but had some oddities:
 * The markup contained multiple `<body>` tags
 * The markup requires a work-around to disable printing
 * The services of `CRM_Core_Resources` aren't reliable within `snippet=2`.
   (This probably depends on which CMS and/or which aspect you're using.
   In Drupal 7, `addVars()` doesn't work with `snippet=2`.)

This commit takes a "back-to-basics" approach.  Since most CMS/Civi services
don't actually make sense on this page, skip them.  Just output the HTML we
want.