CRM_Mosaico_Page_Editor - Generate all mosaico.init() options server-side
In another commit, we'll need to pass some dynamically-generated config data
(i.e.  the token list).  Tangentially, in the current code, we need to
figure out URLs -- but this requires some sketchy bits because it's written
in client-JS.

Since `mosaico.init({...options...})` is really about providing
metadata+policies for the integration, it seems reasonable to compute that
using PHP -- where we have quick access to things like the
`CRM_Utils_System::url()` and `civcrm_api()`.