Navigation Menu - Only show one template management link
At the moment, we store the message templates and mailing templates
separately.  This is not ideal -- the two should be reconciled.
However, in the mean time, listing both at the top-level gets
the user to start-off in a rough place. ("What's the difference
between 'Message Template' and 'Mosaico Template'?")

Ideally, perhaps we would update the `civicrm/a/#/mosaico-template` UI to
allow CRUDing message-templates and deal with msgtpl properties (like
"Subject" and "Text"). That's a bigger project.

This patch is not ideal, but it at least defers the cognitive dissonance.
With this patch, there are two navigation paths to creating Mosaico-based
templates:

 1. When you first install Mosaico, the expectation is to manage
    templates for mail-blasts. You want to get to `civicrm/a/#/mosaico-template`,
    so this is the item in the main-menu.
 2. If you want to send transactional emails using Mosaico-based templates,
    then you're already pretty educated about Civi and have probably seen
    "Administer => Communications => Message Templates"
    (`civicrm/admin/messageTemplates`) before.  Go back to that screen and
    choose the "Mosaico" tab.