Remove 1.x message-template UI from 2.x
There are a few reasons.  The one which spurs action now -- we'd like to try
Mosaio v0.16, and 1.x UI requires forward-porting another conflicted patch
related to `localStorage`.

More broadly, it'll be easier to test/validate the main 2.x functionality if
this stuff isn't in-scope, and IMHO the ideal UI+code for managing
message-templates will require some reworking.

Note: This does *not* affect the data (SQL tables).  Tha information is
still retained in its current form.