crmMosaicoBlockMailing - Fix loading of "Reply To" field
The "Reply To" field is an optional one that appears if you enable
"Administer => CiviMail => Mailer Settings => Enable Custom Reply-To".  If
you set a "Reply To" value, saved, and returned to the mailing, then this
field was not properly redisplayed.

In the original CiviMail code, this field was based on `crm-ui-select` (aka
`select2`).  However, during Mosaico development, there were styling issues
with `crm-ui-select`+Bootstrap, so we downgraded to plain HTML `select`.
Alas, that seems to have created the bug with "Reply To".

Fortunately, `select2`+Bootstrap now looks better than before. So we can
restore to the original, working implementation.

(Note: `select2`+Bootstrap looks *better* but not *great*.  It could be
improved.  But IMHO that's an issue for the themers.)