"Insert Token" Dialog - Adapt size to 50%x66%
Before
------
When inserting a token, the dialog size is hard-coded to `{w:310,h:300}`.
Issue #136 indicates the token dialog is sometimes too small.

After
-----
When inserting a token, the dialog size adapts to `{w:50%,h:66%}` --
with a minimum of 400px in each dimension.

Comments
--------
It would be ideal if the dialog width adapted to the size of the *content*
(eg to handle long token names), but I don't know how to accomplish that
with `editor.windowManager.open()`. And I think this is an improvement anyway:

 * This patch effectively increases the width to at least 500px in the
   average case.  (Rough stat: 95% of laptop/desktop users have a screen
   width of at least 1024px; at full screen, that's 512px+.
   https://www.w3schools.com/browsers/browsers_display.asp)
 * Compared to the hard-coded version, it should scale a bit better to
   different screens/dpis.