info.xml, README.md - Bump official dependency to 5.0
The 4.7 series is pretty long, and the extension is only really compatible
with later versions (mid/high-.20's -- haven't tested exactly).  However,
once we publish the stable release, it'll be presented in-app for all of
them.

This change doesn't make things any better or worse for 4.7 than they are
today. However, it means that (a) the in-app download only appears on 5.0+
and (b) we're not encouraging support/bug-reports from 4.7.