Restore gulp3. Regenerate CSS.
The `gulp` command builds the CSS.  It is currently broken.  The package was
upgraded to gulp4 as part of a long list of updates
(13160c633e12471df7e78a8e3c3998ba1c6a71d6), but the current `gulpfile.js`
only works on gulp3.

Just to demonstrate workingness (and to potentially facilitate debugging if
there are any quirks), I've run `gulp` again to regenerate the CSS.