Fix Leave site? warning when not required
Before: Leave site? warning was triggered by onbeforeunload when saving the mailing draft or submitting the mailing and otherwise leaving the mailing (after the user had edited the design of the mailing).

After: Leave site? warning is only triggered when leaving the mailing while editing the mailing design in Mosaico and not in the CiviCRM parts of the mailing editor.

This change just checks if the Mosaico iframe is shown on the mailing editing page before triggering the Leave site? warning and does not trigger the warning if the iframe is not currently shown.